requirements.txt
Flask==3.0.3
Flask-MySQLdb==2.0.0
mysql-connector-python==9.0.0
fastapi==0.113.0
crewai==0.41.1
crewai_tools==0.4.26
langchain-community==0.2.10
langchain-groq==0.1.9
python-dotenv==1.0.1
groq==0.11.0
streamlit==1.38.0
yfinance==0.2.43
plotly==5.24.0